feat: Add MySQL, PostgreSQL, SQLite, and SQL Server demo files with parameterized queries and formatting examples
